refactor model upload
