+++ deprecated/core/api/upload_model/constants.py
@@ -1,0 +1,4 @@
+import urllib3
+
+
+ORGANISATION_UUID = "61f5a464c35044c19bc7d4b42d7f58cb"  # Nelen & Schuurmans Consultancy
