+++ deprecated/core/api/upload_model/login.py
