+++ hhnk_threedi_tools/external/README.md
@@ -1,0 +1,1 @@
+External functions, keep an eye out that they may be available in a package at some point. This is preferred over including them here.
