+++ tests/test_modelsplitter.py
@@ -5,7 +5,7 @@ import pytest
 
 import hhnk_research_tools as hrt
 from hhnk_threedi_tools.core.folders import Folders
-from hhnk_threedi_tools.core.checks.model_splitter import ModelSchematisations
+from hhnk_threedi_tools.core.schematisation.model_splitter import ModelSchematisations
 from tests.config import FOLDER_TEST, PATH_NEW_FOLDER
